requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0
django-filter==22.1
django-seed==0.3.1
djangorestframework==3.13.1
Faker==13.15.0
generics==3.5.0
on==0.0.4
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toposort==1.7
